% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.weather_function.R
\name{get.weather}
\alias{get.weather}
\title{OpenWeatherMapR}
\usage{
get.weather(key, lat, lon, daterange)
}
\arguments{
\item{key}{string Your Open Weather Map API key.}

\item{location}{City name eg: Manchester.}

\item{date}{Date range: today, yesterday, last5days, "2021-07-01" or list of dates example - list("2021-07,01","2021-07-02")}
}
\value{
Table containing weather information.
}
\description{
Returns weather metrics for specified Location and Date Range from the Open Weather Map API
}
\details{
This is a descriptions
desctiptions
blah blah blah
}
\examples{
\dontrun{
get.weather("APIKEY", Manchester,today)
get.weather("APIKEY", Manchester, "2021-07-01")
}
}
\author{
Chloe Morgan
}
