% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_air_pollution}
\alias{get_air_pollution}
\title{Get Air Pollution Data}
\usage{
get_air_pollution(city)
}
\arguments{
\item{city}{The name of the city.}
}
\value{
A list containing air pollution data.
}
\description{
Retrieves current air pollution data for a given city.
}
