% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_current_weather}
\alias{get_current_weather}
\title{Get Current Weather}
\usage{
get_current_weather(city, units = "metric")
}
\arguments{
\item{city}{The name of the city.}

\item{units}{Units of measurement (default is "metric").}
}
\value{
A list containing current weather data.
}
\description{
Retrieves current weather data for a given city.
}
