% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_daily_forecast}
\alias{get_daily_forecast}
\title{Get Daily Weather Forecast}
\usage{
get_daily_forecast(city, units = "metric")
}
\arguments{
\item{city}{The name of the city.}

\item{units}{Units of measurement (default is "metric").}
}
\value{
A list containing daily weather forecast data.
}
\description{
Retrieves daily weather forecast data for a given city.
}
