% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_hourly_forecast}
\alias{get_hourly_forecast}
\title{Get Hourly Weather Forecast}
\usage{
get_hourly_forecast(city, units = "metric")
}
\arguments{
\item{city}{The name of the city.}

\item{units}{Units of measurement (default is "metric").}
}
\value{
A list containing hourly weather forecast data.
}
\description{
Retrieves hourly weather forecast data for a given city.
}
