% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_location}
\alias{get_location}
\title{Get City Coordinates}
\usage{
get_location(city)
}
\arguments{
\item{city}{The name of the city.}
}
\value{
A list containing the latitude and longitude of the city.
}
\description{
Retrieves the latitude and longitude of a given city.
}
