% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_multiple_cities_weather}
\alias{get_multiple_cities_weather}
\title{Get Weather Data for Multiple Cities}
\usage{
get_multiple_cities_weather(cities, units = "metric")
}
\arguments{
\item{cities}{A vector of city names.}

\item{units}{Units of measurement (default is "metric").}
}
\value{
A list containing weather data for multiple cities.
}
\description{
Retrieves current weather data for multiple cities.
}
