% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_weather_by_coordinates}
\alias{get_weather_by_coordinates}
\title{Get Current Weather by Coordinates}
\usage{
get_weather_by_coordinates(lat, lon, units = "metric")
}
\arguments{
\item{lat}{The latitude.}

\item{lon}{The longitude.}

\item{units}{Units of measurement (default is "metric").}
}
\value{
A list containing current weather data.
}
\description{
Retrieves current weather data for a given set of coordinates.
}
