% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_api.R
\name{get_weather_forecast}
\alias{get_weather_forecast}
\title{Get Weather Forecast}
\usage{
get_weather_forecast(city, units = "metric")
}
\arguments{
\item{city}{The name of the city.}

\item{units}{Units of measurement (default is "metric").}
}
\value{
A list containing weather forecast data.
}
\description{
Retrieves weather forecast data for a given city.
}
