% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set OpenWeather API Key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{Your OpenWeather API key.}
}
\description{
Set your OpenWeather API key as an environment variable.
}
