% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.weather_function.R
\name{setup.weather}
\alias{setup.weather}
\title{OpenWeatherMapR}
\usage{
setup.weather(key, location, date)
}
\arguments{
\item{key}{string Your Open Weather Map API key.}

\item{location}{City name eg: Manchester.}

\item{date}{Date range: today, yesterday, last5days, "2021-07-01"}
}
\value{
Table containing weather information.
}
\description{
Returns weather metrics for specified Location and Date Range from the Open Weather Map API
}
\details{
This is a descriptions
desctiptions
blah blah blah
}
\examples{
\dontrun{
setup.weather("APIKEY", Manchester,today)
setup.weather("APIKEY", Manchester, "2021-07-01")
}
}
\author{
Chloe Morgan
}
